requirements.txt
alpaca-py>=0.25.0
pandas>=2.0.0
python-dotenv>=1.0.0
pytz>=2023.3
pytest>=7.4.0
